dailybuild: trivial warnings の抑止（src/Samples)
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6791 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Samples/Physics/Joints/main.cpp
@@ -91,7 +91,7 @@ public:
 		handlers.push_back(DBG_NEW ChainHandler(this));
 		handlers.push_back(DBG_NEW LinkHandler(this));
 		handlers.push_back(DBG_NEW ArmHandler(this));
-		numScenes	= handlers.size();
+		numScenes	= (int)handlers.size();
 	}
 	virtual ~MyApp(){}
